backend/requirements.txt
fastapi
uvicorn
ollama
pydantic
pytest
httpx
Pillow
TTS
python-multipart
openai-whisper
google-adk
litellm